users/views.py
from django.shortcuts import render

# Create your views here.
from rest_framework import generics
from .serializers import UserRegisterSerializer
from django.contrib.auth import get_user_model

User = get_user_model()


class UserRegisterView(generics.CreateAPIView):
    queryset = User.objects.all()
    serializer_class = UserRegisterSerializer
