requirements.txt
# === Основные зависимости ===
Django==5.0.6
djangorestframework==3.15.2
django-filter==24.2
django-cors-headers==4.3.1

# === Аутентификация и токены ===
djangorestframework-simplejwt==5.3.1

# === Асинхронные задачи и брокер ===
celery==5.4.0
redis==5.0.4

# === Telegram интеграция ===
python-telegram-bot==21.4

# === Тестирование и качество кода ===
coverage==7.5.3
flake8==7.1.1

# === Дополнительно ===
python-dotenv==1.0.1
requests==2.32.3
